+++ py_abac/storage/file/__init__.py
@@ -1,3 +1,5 @@
 """
-    File storage
+    Exposed classes and methods
 """
+
+from .storage import FileStorage