+++ CHANGELOG.md
@@ -29,4 +29,4 @@
 - Added `RedisStorage` backend.
 - Added `FileStorage` backend.
 - Fixed typos in documentation.
-- Fixed import dependency error for storage. Removed import statements from `py_abac/storage/__init__.py`.
+- Removed all import statements from `py_abac/storage/__init__.py`.